Agent1/requirements.txt
google-generativeai>=0.7.0
chromadb>=0.4.0
openai-whisper>=20231117
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
